requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
python-dotenv
email-validator
flask-bcrypt
Pillow
flask_login
flask_bcrypt